Difierenciando Mock e Stub
