+++ src/CursoOnline.Dominio/Cursos/CursoDto.cs
@@ -1,0 +1,15 @@
+﻿namespace CursoOnline.Dominio.Cursos
+{
+    public class CursoDto
+    {
+        public CursoDto()
+        {
+        }
+
+        public string Nome { get; set; }
+        public string Descricao { get; set; }
+        public double CargaHoraria { get; set; }
+        public string PublicoAlvo { get; set; }
+        public double Valor { get; set; }
+    }
+}
