+++ src/CursoOnline.Dominio/Cursos/ICursoRepositorio.cs
@@ -1,0 +1,8 @@
+﻿namespace CursoOnline.Dominio.Cursos
+{
+    public interface ICursoRepositorio
+    {
+        void Adicionar(Curso curso);
+        Curso ObterPeloNome(string nome);
+    }
+}
